Refactoring what controllers are available.
