+++ src/Controllers/AbstractFileUploadController.php
@@ -1,0 +1,20 @@
+<?php
+
+declare(strict_types=1);
+
+namespace ResourceController\Controllers;
+
+use ResourceController\Traits\Controllers\AssignOwner;
+use ResourceController\Traits\Controllers\UploadMediaFile;
+
+abstract class AbstractFileUploadController extends AbstractModelController
+{
+    use UploadMediaFile, AssignOwner;
+
+    /**
+     * @var array
+     */
+    protected $skipAttributes = [
+        'media',
+    ];
+}
