+++ src/Controllers/JsonHtml/UserController.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace ResourceController\Controllers\Html;
+namespace ResourceController\Controllers\JsonHtml;
 
 use Illuminate\Support\Facades\Hash;
 use Illuminate\Foundation\Http\FormRequest;
