+++ src/Controllers/UserController.php
@@ -1,0 +1,28 @@
+<?php
+
+declare(strict_types=1);
+
+namespace ResourceController\Controllers;
+
+use Illuminate\Support\Facades\Hash;
+use Illuminate\Foundation\Http\FormRequest;
+
+abstract class UserController extends AbstractModelController
+{
+    /**
+     * Get attributes from validated request.
+     *
+     * @param  \Illuminate\Foundation\Http\FormRequest  $request
+     * @return array
+     */
+    protected function getRequestAttributes(FormRequest $request): array
+    {
+        $attributes = parent::getRequestAttributes($request);
+
+        if (isset($attributes['password'])) {
+            $attributes['password'] = Hash::make($attributes['password']);
+        }
+
+        return $attributes;
+    }
+}
